+++ install/empty_data.php
@@ -8049,7 +8049,7 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_SUPER_ADMIN,
                 'name' => 'rule_change',
-                'rights' => READ | UPDATE | CREATE | PURGE | RuleTicket::PARENT,
+                'rights' => READ | UPDATE | CREATE | PURGE | RuleCommonITILObject::PARENT,
             ], [
                 'profiles_id' => self::PROFILE_ADMIN,
                 'name' => 'rule_change',
@@ -8057,6 +8057,18 @@ style="color: #8b8c8f; font-weight: bold; text-decoration: underline;"&gt;
             ], [
                 'profiles_id' => self::PROFILE_READ_ONLY,
                 'name' => 'rule_change',
+                'rights' => READ,
+            ], [
+                'profiles_id' => self::PROFILE_SUPER_ADMIN,
+                'name' => 'rule_problem',
+                'rights' => READ | UPDATE | CREATE | PURGE | RuleCommonITILObject::PARENT,
+            ], [
+                'profiles_id' => self::PROFILE_ADMIN,
+                'name' => 'rule_problem',
+                'rights' => READ,
+            ], [
+                'profiles_id' => self::PROFILE_READ_ONLY,
+                'name' => 'rule_problem',
                 'rights' => READ,
             ],
             [